Updating code to have the menu that was written. Changed the name of the java classes and xml files to reflect their functions. This did involve deleting/replacing some files so hopefully the files will be updated appropriately on github. The sixteen lights code is there, but it doesn't seem to be accessible at the moment. Also, lots of thing will cause the app to crash. I'm not sure if this is since I've changed anything or if it was like this when the code had been run in anyone else's project.
